In progress get-campaigns-reports refactor level-2
