+++ src/routes/campaigns/campaignId/reports/_get/index.ts
@@ -1,53 +1,18 @@
-/** OPENAPI-ROUTE: get-campaigns-reports */
-import { Context } from "openapi-backend";
-import { ERROR_MESSAGE } from "@src/utils/constants";
+/** OPENAPI-CLASS: get-campaigns-reports */
+import CampaignRoute from "@src/features/routes/CampaignRoute";
 import { getCampaign, getCampaignReports } from "@src/utils/campaigns";
-import { getProjectById } from "@src/utils/projects";
 
-export default async (
-  c: Context,
-  req: OpenapiRequest,
-  res: OpenapiResponse
-) => {
-  let user = req.user;
-  let error = {
-    code: 500,
-    message: ERROR_MESSAGE,
-    error: true,
-  } as StoplightComponents["schemas"]["Error"];
+export default class Route extends CampaignRoute<{
+  response: StoplightOperations["get-campaigns-cid-meta"]["responses"]["200"]["content"]["application/json"];
+  parameters: StoplightOperations["get-campaigns-cid-meta"]["parameters"]["path"];
+}> {
+  protected async prepare() {
+    if (!(await getCampaign({ campaignId: this.cp_id })))
+      return this.setError(403, {} as OpenapiError);
 
-  let cid = parseInt(c.request.params.cid as string);
+    const reports: any = await getCampaignReports(this.cp_id);
+    if (!reports) return this.setError(403, {} as OpenapiError);
 
-  res.status_code = 200;
-
-  try {
-    // Check if the campaign exists
-    let campaign = await getCampaign({ campaignId: cid });
-
-    if (!campaign) {
-      error.code = 400;
-      throw error;
-    }
-
-    // Check if user has permission to edit the campaign
-    await getProjectById({
-      projectId: campaign.project.id,
-      user: user,
-    });
-
-    // Get the campaign reports
-    const reports = await getCampaignReports(cid);
-
-    return reports as StoplightComponents["schemas"]["Report"][];
-  } catch (e: any) {
-    if (e.code) {
-      error.code = e.code;
-      res.status_code = e.code;
-    } else {
-      error.code = 500;
-      res.status_code = 500;
-    }
-
-    return error;
+    return this.setSuccess(200, reports);
   }
-};
+}
